Reworked file hierarchy, added basic terrain gen Fully restructured the file hierarchy. Moved the JSON files to /resources/json Moved the python files (trigger, property, and placements) that should be extended to their own directory (/extendable_python/...) These will probably need to import all local files eventually, to enable extensions. Broke all the imports. Fixed all the imports. Added a lot of __init__ files to make relative imports work properly. (This was the least hacky way to make imports work with this hierarchy.) (Thanks, Python import design)
Renamed a lot of stuff.
Made docstrings google-style.
Fixed a bunch of indentation for multiline statements to be inline with PEP8.

Added a very small amount of error handling:
Property functions that cause an exception now default to the Value field
Placement functions default to 0
Both types raise a warning

Added a basic simplex noise city terrain generation
We might want to move the actual generation bit around later
Although as-is might be okay too, given that it makes it easy to change the algorithm