I should have written this commit message earlier. Reconfigured extensible python file structure New structure: All function classes executable by extensible components inherit from ComponentFunction This includes triggered, threshold, and placement functions Each class can be accessed from ComponentFunctionManager.componentFunctions (dict {str: class}) The callable functions all use a .execute method, with standardized args The ComponentFunctionManager imports every file in the top layer of the extensible_python submodule (See /extensible_python/__init__.py for details) Basic functions should reside in this module Ones with more dependencies should import further sub-submodules (These should, of course, reside in /extensible_python)
TerrainComponents now have a terrainObject property
This should hold an object representing the terrain in an appropriate coordinate system,
generated with an appropriate algorithm by a trigger
An abstract base class for TerrainObjects will need to be created later
(Though actual implementations should still live in /extensible_python/terrain)

Also created component_exceptions, which should probably be renamed later
(to something more general)

Pulled placeObject abstract method and default call from MapComponent
This can live in subclassInit for MapPoint
(As it's not really general enough to belong to every subclass).

Finally, I'm putting development of this branch on hold for a while
While I intend on completing it Later, for now
I'd prefer to get a functional version of a less extensible variant built
There will be a lot of work to make this version efficient enough to be usable
Note: Probably going to need a time manager object
(Similar to LayerManager? Related? TBD.)

I've probably forgotten a few changes, but that should be most.